requirements.txt
Django==4.2.7
psycopg2-binary==2.9.7
django-cors-headers==4.3.1
Pillow==10.0.1
gunicorn==21.2.0
django-widget-tweaks==1.4.12
weasyprint==62.1
requests==2.31.0
python-decouple==3.8
redis==5.0.1
django-redis==5.4.0
celery==5.3.4
python-dotenv==1.0.0